now that I'm tinkering with bridging, the pfil hook ends up doing IP interception on both the incoming physical interface (netmap0/netmap1 depending upon direction) and bridge0.
Now, to add annoyance to annoyance, IFF_PROMISCINET is only set on the
interface we are doing promiscuous passive ip intereption on.
In this instance it's bridge0.  So the frames from netmapX don't have l2
information.

So, extend the API a little more so i can tell it which ifnet to sniff for.
Later on I'll re-extend it to pass everything up and filter on the userland
code side.  ight now though it's hard to do as the ifnet doesn't have
access to the cfg bits in any generic way - ifnet->ifp_softc points to the
interface specific type, not the generic config section.

Anyway.  This'll do for now.